Add files via upload
This one includes the version of NEQAIR input where and LOS file can be read for species information, bands can be selected based on the species, and some of the new NEQAIR.inp file can be written out.. Still needs to add functionality for including band systems for each species, and then move on to other aspects of input file. Most likely creating two different tabs for the inputs and then splitting them that way.